Rename utils.go to parser.go & Fixes
